+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - **Feature:** Add new field `AltPort` to `ActiveHealthCheck`
   - **Feature:** Add new field `Tls` to `HttpHealthCheck`
   - **Breaking change:** Renamed `TargetPoolTlsConfig` to `TlsConfig`
+- `loadbalancer`: [v1.8.0](services/loadbalancer/CHANGELOG.md#v180)
+  - **Feature:** Add new fields `AltPort` and `HttpHealthCheck` to `ActiveHealthCheck`
 
 ## Release (2026-02-20)
 - `core`: [v0.21.1](core/CHANGELOG.md#v0211)